Fixed the api data access
